pyproject.toml
[tool.poetry]
name = "layton-eval"
version = "0.1.0"
description = ""
authors = ["Raphael Vienne <raphael.vienne@live.fr>"]
readme = "README.md"
packages = [{include = "layton_eval", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
numpy = "^1.26.3"
tqdm = "^4.66.1"
pillow = "^10.2.0"
pandas = "^2.1.4"
openpyxl = "^3.1.2"
xlsxwriter = "^3.1.9"
pyxlsb = "^1.0.10"
langchain = "^0.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
ruff = "^0.1.10"
pytest = "^7.4.4"
pytest-sugar = "^0.9.7"
pre-commit = "^3.6.0"
coverage = "^7.4.0"
ipykernel = "^6.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.black]
skip-string-normalization = true
line-length = 120

[tool.ruff]
# Same as Black.
line-length = 120

exclude = ["jupyter_notebook_config.py"]
select = [
    "E",  # pycodestyle errors (settings from FastAPI, thanks, @tiangolo!)
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "C901",  # too complex
]

[tool.ruff.isort]
order-by-type = true
relative-imports-order = "closest-to-furthest"
extra-standard-library = ["typing"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
known-first-party = []
